pyproject.toml
[project]
name = "router-architecture"
version = "0.1.0"
description = "Production-ready multi-model LLM routing system"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
router = "router.cli:cli"
router-api = "router.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/router"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: fast, no external dependencies",
    "integration: requires live provider endpoints",
    "calibration: runs all providers against calibration prompts",
]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]
